moved menubar css to own css file, updated imports
